Rename page_input_providers.py to providers.py
