+++ tests/test_utils.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 import scrapy
 from scrapy.http import TextResponse
 
-from scrapy_poet.page_input_providers import (
+from scrapy_poet.providers import (
     PageObjectInputProvider,
     ResponseDataProvider,
     provides,
